+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
  - Added support for `pass` statement on `class` level
+ - Added a possibility to disable linter checks in range from e.g. `# gdlint: disable=function-name` to `# gdlint: enable=function-name`
 
 ### Fixed
  - Fixed comment persistence check
@@ -18,7 +19,7 @@
  - Added support for legacy (godot `3.1`) parenthesesless `assert` (e.g. `assert 1 == 1`)
  - Added support for recursive file finding in `gdlint` (e.g. `gdlint dir1/ dir2/`)
  - Added pre-commit hooks which can consumed by `pre-commit` project
- - Added a possibility to disable linter checks (e.g. `# gdlint: ignore=function-name` inline or in the line before)
+ - Added a possibility to ignore linter checks inplace (e.g. `# gdlint: ignore=function-name` inline or in the line before)
  - Added support for underscored numeric literals such as `1_000_000`, `1_000.000`, and `0xff_99_00`
 
 ### Fixed